Minor fix for time range entry display on form

+++ CoffeeNewspaper/CN_Presentation/ViewModel/Input/Interfaces/IUpdateTimeRange.cs
@@ -4,6 +4,6 @@ namespace CN_Presentation.Input
 {
     public interface IUpdateTimeRange
     {
-        void NotifyUpdateTimeRange(int timeRangeSeconds);
+        void NotifyUpdateTimeRange(long timeRangeSeconds);
     }
 }
